created new controllers for each game type
